requirements.txt
requests~=2.0
pytest~=8.1
python-dotenv~=1.0.1
